+++ main/app/api/admin/quizes/bulk-actions/delete-selected/route.tsx
@@ -1,36 +1,82 @@
 import prisma from "@/app/libs/db";
 import { NextResponse } from "next/server";
+import { type NextRequest } from 'next/server';
+import jwt from "jsonwebtoken";
 
 interface Respo {
     success: boolean,
     message: string
 }
 
-export async function DELETE(req: Request) {
+export async function DELETE(req: NextRequest) {
     let resp: Respo = {
         success: false,
         message: ''
     }
 
-    let sts: number = 400;
+    let sts: number = 200;
+    let isTrueAdminUser: boolean = false;
 
     try {
 
         const body = await req.json();
-        const { quiz_id_list } = body;
+        const { token, quiz_id_list } = body;
 
-        if (quiz_id_list) {
-            await prisma.qF_Quiz.deleteMany({
-                where: {
-                    quiz_id: {
-                        in: quiz_id_list
+        if (token && quiz_id_list) {
+            const res = jwt.verify(token as string, process.env.JWT_SECRET ?? "") as { is_admin_user: string };
+
+            if (res) {
+
+                const user_id = res.is_admin_user;
+
+                const fu__in__usrtblmdl = await prisma.qF_User.findFirst({
+                    where: {
+                        AND: [
+                            {
+                                user_id
+                            },
+                            {
+                                role: "Admin"
+                            }
+                        ]
+                    }
+                });
+                const fu__in__admntblmdl = await prisma.qF_Admin_User.findFirst({
+                    where: {
+                        admin_user_id: user_id,
+                    }
+                });
+
+                if (fu__in__usrtblmdl) {
+                    isTrueAdminUser = true;
+                } else {
+                    if (fu__in__admntblmdl) {
+                        isTrueAdminUser = true;
+                    } else {
+                        isTrueAdminUser = false;
                     }
                 }
-            });
-            sts = 200;
-            resp = {
-                success: true,
-                message: "Selected Quizes Deleted Successfully!"
+
+                if (isTrueAdminUser) {
+                    await prisma.qF_Quiz.deleteMany({
+                        where: {
+                            quiz_id: {
+                                in: quiz_id_list
+                            }
+                        }
+                    });
+                    sts = 200;
+                    resp = {
+                        success: true,
+                        message: "Selected Quizes Deleted Successfully!"
+                    }
+                } else {
+                    resp = {
+                        success: false,
+                        message: 'User Not Found.',
+                    }
+                    sts = 200;
+                }
             }
         } else {
             sts = 400;
@@ -43,10 +89,37 @@ export async function DELETE(req: Request) {
         return NextResponse.json(resp, { status: sts });
         //eslint-disable-next-line
     } catch (error: any) {
-        sts = 500;
-        resp = {
-            success: false,
-            message: error.message
+        // sts = 500;
+        // resp = {
+        //     success: false,
+        //     message: error.message
+        // }
+        if (error.message == "jwt expired") {
+            resp = {
+                success: false,
+                message: "Your session is expired, Please login again."
+            }
+        } else if (error.message == "jwt malformed" || error.message == "jwt must be a string") {
+            resp = {
+                success: false,
+                message: "Wrong information provided."
+            }
+        } else if (error.message == "invalid signature" || error.message == "invalid token") {
+            resp = {
+                success: false,
+                message: "Invalid information provided."
+            }
+        } else if (error.message == "jwt must be provided") {
+            sts = 400;
+            resp = {
+                success: false,
+                message: "Missing required fields."
+            }
+        } else {
+            resp = {
+                success: false,
+                message: error.message
+            }
         }
         return NextResponse.json(resp, { status: sts });
     }
