+++ main/app/api/admin/quizes/crud/delete/route.tsx
@@ -1,5 +1,6 @@
 import prisma from "@/app/libs/db";
 import { NextResponse } from "next/server";
+import jwt from "jsonwebtoken";
 
 interface Respo {
     success: boolean,
@@ -12,35 +13,80 @@ export async function DELETE(req: Request) {
         message: ''
     }
 
-    let sts: number = 400;
+    let sts: number = 200;
+    let isTrueAdminUser: boolean = false;
 
     try {
 
         const body = await req.json();
-        const { quiz_id } = body;
+        const { token, quiz_id } = body;
 
-        if (quiz_id) {
-            const alreadQuizExited = await prisma.qF_Quiz.findFirst({
-                where: {
-                    quiz_id
-                }
-            });
-            if (alreadQuizExited) {
-                await prisma.qF_Quiz.delete({
+        if (token && quiz_id) {
+
+            const res = jwt.verify(token as string, process.env.JWT_SECRET ?? "") as { is_admin_user: string };
+
+            if (res) {
+
+                const user_id = res.is_admin_user;
+
+                const fu__in__usrtblmdl = await prisma.qF_User.findFirst({
                     where: {
-                        quiz_id
+                        AND: [
+                            {
+                                user_id
+                            },
+                            {
+                                role: "Admin"
+                            }
+                        ]
                     }
                 });
-                sts = 200;
-                resp = {
-                    success: true,
-                    message: "Quiz Deleted Successfully!"
+                const fu__in__admntblmdl = await prisma.qF_Admin_User.findFirst({
+                    where: {
+                        admin_user_id: user_id,
+                    }
+                });
+
+                if (fu__in__usrtblmdl) {
+                    isTrueAdminUser = true;
+                } else {
+                    if (fu__in__admntblmdl) {
+                        isTrueAdminUser = true;
+                    } else {
+                        isTrueAdminUser = false;
+                    }
                 }
-            } else {
-                sts = 200;
-                resp = {
-                    success: false,
-                    message: "Quiz Not Exist!"
+
+                if (isTrueAdminUser) {
+                    const alreadQuizExited = await prisma.qF_Quiz.findFirst({
+                        where: {
+                            quiz_id
+                        }
+                    });
+                    if (alreadQuizExited) {
+                        await prisma.qF_Quiz.delete({
+                            where: {
+                                quiz_id
+                            }
+                        });
+                        sts = 200;
+                        resp = {
+                            success: true,
+                            message: "Quiz Deleted Successfully!"
+                        }
+                    } else {
+                        sts = 200;
+                        resp = {
+                            success: false,
+                            message: "Quiz Not Exist!"
+                        }
+                    }
+                } else {
+                    resp = {
+                        success: false,
+                        message: 'User Not Found.',
+                    }
+                    sts = 200;
                 }
             }
         } else {
@@ -61,10 +107,37 @@ export async function DELETE(req: Request) {
                 message: "Quiz Not Exist!"
             }
         } else {
-            sts = 500;
-            resp = {
-                success: false,
-                message: error.message
+            // sts = 500;
+            // resp = {
+            //     success: false,
+            //     message: error.message
+            // }
+            if (error.message == "jwt expired") {
+                resp = {
+                    success: false,
+                    message: "Your session is expired, Please login again."
+                }
+            } else if (error.message == "jwt malformed" || error.message == "jwt must be a string") {
+                resp = {
+                    success: false,
+                    message: "Wrong information provided."
+                }
+            } else if (error.message == "invalid signature" || error.message == "invalid token") {
+                resp = {
+                    success: false,
+                    message: "Invalid information provided."
+                }
+            } else if (error.message == "jwt must be provided") {
+                sts = 400;
+                resp = {
+                    success: false,
+                    message: "Missing required fields."
+                }
+            } else {
+                resp = {
+                    success: false,
+                    message: error.message
+                }
             }
         }
         return NextResponse.json(resp, { status: sts });
